pyproject.toml
[tool.poetry]
name = "NexusLM"
version = "0.1.0"
description = "customer service AI Agent using Agent Development Kit"
authors = ["Cedric Sebastian"]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-adk = "^1.0.0"  # No newer version available
jsonschema = "^4.23.0"
hubspot-api-client = "^1.0.0"  # Added new dependency
google-cloud-aiplatform = { extras = ["adk", "agent_engine", "evaluation"], version = "^1.100.0" }

[tool.poetry.group.dev.dependencies]
pytest = "^8.12.0"
pytest-mock = "^3.11.1"
scikit-learn = "^1.3.2"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"
flake8-pyproject = "^1.2.3"  # No newer version available
pylint = "^3.10.0"
pyink = "^24.10.1"  # No newer version available
jsonschema = "^4.19.3"
# google-cloud-aiplatform = { extras = ["evaluation"], version = "^1.93.0" }  # Removed redundant entry


[tool.pytest.ini_options]
console_output_style = "progress"
addopts = "-vv -s"
#addopts = "-vv -s --pdb"
testpaths = ["tests/", "eval/"]
markers = ["unit"]
log_level = "ERROR"
log_cli = false
log_auto_indent = true
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_cli_format = "[%(asctime)s] %(levelname)s (%(funcName)s) \t [%(pathname)s:%(lineno)d] %(message)s"
filterwarnings = ["ignore::UserWarning"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.mason"


[tool.pyink]
line-length = 80
pyink-indentation = 4
pyink-use-majority-quotes = true
